Factor out HTTP client to a separate module
